Refactored Data
