+++ src/main/java/rs117/hd/HdPlugin.java
@@ -104,10 +104,10 @@ import rs117.hd.config.AntiAliasingMode;
 import rs117.hd.config.DefaultSkyColor;
 import rs117.hd.config.FogDepthMode;
 import rs117.hd.config.UIScalingMode;
-import rs117.hd.environments.EnvironmentManager;
-import rs117.hd.lighting.LightManager;
-import rs117.hd.lighting.SceneLight;
-import rs117.hd.materials.Material;
+import rs117.hd.scene.EnvironmentManager;
+import rs117.hd.scene.lighting.LightManager;
+import rs117.hd.scene.lighting.SceneLight;
+import rs117.hd.data.materials.Material;
 import rs117.hd.model.TempModelInfo;
 import rs117.hd.model.objects.ObjectProperties;
 import rs117.hd.model.ModelHasher;
@@ -126,7 +126,7 @@ import static org.jocl.CL.clCreateFromGLBuffer;
 
 import rs117.hd.scene.ProceduralGenerator;
 import rs117.hd.scene.SceneUploader;
-import rs117.hd.scene.data.TextureManager;
+import rs117.hd.scene.TextureManager;
 import rs117.hd.utils.*;
 import rs117.hd.utils.buffer.GLBuffer;
 import rs117.hd.utils.buffer.GpuFloatBuffer;
