+++ src/main/java/rs117/hd/data/environments/Area.java
@@ -22,10 +22,11 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package rs117.hd.environments;
+package rs117.hd.data.environments;
 
 import lombok.Getter;
 import net.runelite.api.Constants;
+import rs117.hd.utils.Rect;
 
 @Getter
 public enum Area
