+++ src/main/java/rs117/hd/data/materials/GroundMaterial.java
@@ -22,7 +22,7 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package rs117.hd.materials;
+package rs117.hd.data.materials;
 
 import java.util.Random;
 import lombok.Getter;
