+++ src/main/java/rs117/hd/model/ModelPusher.java
@@ -6,11 +6,14 @@ import lombok.extern.slf4j.Slf4j;
 import net.runelite.api.*;
 import rs117.hd.HdPlugin;
 import rs117.hd.HdPluginConfig;
+import rs117.hd.data.materials.Material;
+import rs117.hd.data.materials.Overlay;
+import rs117.hd.data.materials.Underlay;
+import rs117.hd.data.materials.UvType;
 import rs117.hd.model.objects.ObjectProperties;
 import rs117.hd.model.objects.ObjectType;
 import rs117.hd.scene.ProceduralGenerator;
-import rs117.hd.lighting.BakedModels;
-import rs117.hd.materials.*;
+import rs117.hd.data.BakedModels;
 import rs117.hd.model.objects.TzHaarRecolorType;
 import rs117.hd.utils.HDUtils;
 import rs117.hd.utils.buffer.FixedLengthHashCode;
