+++ src/main/java/rs117/hd/scene/TextureManager.java
@@ -22,7 +22,7 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package rs117.hd.scene.data;
+package rs117.hd.scene;
 
 import com.jogamp.opengl.GL4;
 import java.awt.image.BufferedImage;
