+++ src/main/resources/rs117/hd/scene/lighting/lights.json
